- Revert last two changes, they weren't very good fixes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/release/V1.2@27789 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csgeom/solidspace.cpp
@@ -169,7 +169,7 @@ void csSolidSpace::AddClosedObject (iTriangleMesh* object)
   csTriangleMeshTools::SortTrianglesX (object, tris, tri_count, planes);
 
   CheckBox (root_bbox, root, object->GetVertices (),
-  	tris, (int)tri_count, planes);
+  	tris, tri_count, planes);
 
   delete[] tris;
   delete[] planes;
